requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.17.1
funkify==0.3.0
geohash2==1.1
googletrans==3.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.21.2
psycopg2==2.9.3
pycodestyle==2.8.0
pygeohash==1.2.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
requires==0.9.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
